requirements.txt
django-markdown-deux
Django>=1.6.2
django-allauth>=0.16.1
django-simple-history
South
diff-match-patch
django-voting
djangorestframework
markdown>=2.1.0
django-filter>=0.5.4
defusedxml>=0.3
